Refactor namespaces and improve async handling
- Changed namespaces for several files to better organize code.
- Added using directive for `DigitalRightsManagement.Infrastructure.Persistence.DbManagement` in `ApiFixture`.
- Updated `BasicAuthHandler` to use `ConfigureAwait(false)` for better async practices.
- Modified `ProductTests` and `UserTests` to include new using directive for shared abstractions.
- Adjusted overall structure of test files for clarity and maintainability.
